message_list_view: Move relevant types to module.
